+++ src/redturtle/chefcookie/browser/viewlets.py
@@ -4,8 +4,13 @@ from plone.app.layout.viewlets.common import ViewletBase
 from plone.memoize import ram
 from redturtle.chefcookie.interfaces import IChefCookieSettings
 from time import time
-
 import pkg_resources
+import six
+
+if six.PY2:
+    from urlparse import urlparse
+else:
+    from urllib.parse import urlparse
 
 
 CHEFCOOKIE_URL = "{portal_url}/++resource++redturtle.chefcookie/chefcookie/chefcookie.min.js?v={version}"
@@ -22,10 +27,25 @@ class GetChefcookieJs(ViewletBase):
     def get_version(self):
         return pkg_resources.get_distribution("redturtle.chefcookie").version
 
+    def domain_allowed(self, domain_whitelist, current_url):
+
+        if not filter(bool, domain_whitelist):
+            return True
+        for domain in domain_whitelist:
+            if domain in current_url:
+                return True
+        return False
+
     def have_chefcookie_configuration(self):
         try:
-            self.context.portal_registry.forInterface(IChefCookieSettings)
-            return True
+            cc = self.context.portal_registry.forInterface(IChefCookieSettings)
+
+            if cc.enable_cc and self.domain_allowed(
+                cc.domain_whitelist, urlparse(self.request.get("URL")).netloc
+            ):
+                return True
+
+            return False
         except Exception:
             return False
 
@@ -39,9 +59,7 @@ class GetChefcookieJs(ViewletBase):
         return (
             CHEFCOOKIE_URL.format(portal_url=portal_url, version=version),
             RT_CHEFCOOKIE_URL.format(
-                portal_url=portal_url,
-                type=self.get_type(),
-                version=version,
+                portal_url=portal_url, type=self.get_type(), version=version,
             ),
             CONFIG_URL.format(
                 portal_url=portal_url,
@@ -52,7 +70,8 @@ class GetChefcookieJs(ViewletBase):
 
     def get_css_link(self):
         return "{portal_url}/++resource++redturtle.chefcookie/styles.css?v={version}".format(
-            portal_url=api.portal.get().portal_url(), version=self.get_version()
+            portal_url=api.portal.get().portal_url(),
+            version=self.get_version(),
         )
 
     def get_type(self):